# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Listing {
  id: string;
  name: string;
  city: string;
  country: string;
  type: string;
  description: string;
  imageUrl: string;
  pricePerNight: number;
  rating: number;
  reviewCount: number;
  amenities: string[];
  maxGuests: number;
  roomLabel: string;
}

export type AssistantMessageInputContext = { [key: string]: unknown };

export interface AssistantMessageInput {
  /** @minLength 1 */
  message: string;
  context?: AssistantMessageInputContext;
}

export type AssistantReplyExtracted = { [key: string]: unknown };

export interface AssistantReply {
  message: string;
  intent: string;
  suggestedListings?: Listing[];
  extracted?: AssistantReplyExtracted;
}

export interface BookingInput {
  listingId: string;
  checkIn: string;
  checkOut: string;
  /** @minimum 1 */
  guests: number;
  /** @minLength 2 */
  customerName: string;
  /** @minLength 3 */
  customerEmail: string;
  /** @minLength 7 */
  customerPhone: string;
  addOns: string[];
}

export type Booking = BookingInput & {
  reference: string;
  listing: Listing;
  nights: number;
  subtotal: number;
  taxes: number;
  fees: number;
  total: number;
  status: string;
  createdAt: string;
};

export interface ErrorResponse {
  error: string;
}

export type GetListingsParams = {
destination?: string;
checkIn?: string;
checkOut?: string;
/**
 * @minimum 1
 */
guests?: number;
/**
 * @minimum 0
 */
maxPrice?: number;
amenity?: string;
};