# lib/api-zod/src/generated/types/assistantMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantMessageInputContext } from './assistantMessageInputContext';

export interface AssistantMessageInput {
  /** @minLength 1 */
  message: string;
  context?: AssistantMessageInputContext;
}