# lib/api-zod/src/generated/types/assistantReply.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantReplyExtracted } from './assistantReplyExtracted';
import type { Listing } from './listing';

export interface AssistantReply {
  message: string;
  intent: string;
  suggestedListings?: Listing[];
  extracted?: AssistantReplyExtracted;
}