# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInput } from './bookingInput';
import type { Listing } from './listing';

export type Booking = BookingInput & {
  reference: string;
  listing: Listing;
  nights: number;
  subtotal: number;
  taxes: number;
  fees: number;
  total: number;
  status: string;
  createdAt: string;
};