# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  listingId: string;
  checkIn: string;
  checkOut: string;
  /** @minimum 1 */
  guests: number;
  /** @minLength 2 */
  customerName: string;
  /** @minLength 3 */
  customerEmail: string;
  /** @minLength 7 */
  customerPhone: string;
  addOns: string[];
}