# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Listing {
  id: string;
  name: string;
  city: string;
  country: string;
  type: string;
  description: string;
  imageUrl: string;
  pricePerNight: number;
  rating: number;
  reviewCount: number;
  amenities: string[];
  maxGuests: number;
  roomLabel: string;
}